feat(ui): overlay titlebar and header panel toggle icons
Remove the native macOS titlebar by switching to overlay mode,
making the app header the topmost UI element. Move panel toggle
buttons (sidebar, terminal, changes) from the status bar into
each view's header using lucide PanelLeft/Bottom/Right icons
with filled active state, matching VS Code's layout pattern.

- Set titleBarStyle: Overlay + hiddenTitle in Tauri window config
- Add data-tauri-drag-region to all view headers for window dragging
- Add macOS traffic-light padding (78px left, 38px top for sidebars)
- Create shared PanelToggles component with SVG rect fill on active
- Embed PanelToggles in ChatPanel, DiffViewer, and Dashboard headers
- Strip toggle buttons from StatusBar, keeping only stats
