+++ src-tauri/tauri.conf.json
@@ -14,7 +14,8 @@
         "width": 1200,
         "height": 800,
         "center": true,
-        "decorations": true
+        "titleBarStyle": "Overlay",
+        "hiddenTitle": true
       }
     ],
     "security": {
