+++ src/ui/src/components/chat/ChatPanel.module.css
@@ -14,6 +14,10 @@
   box-shadow: var(--shadow-sm);
 }
 
+[data-platform="mac"] .header {
+  padding-left: 78px;
+}
+
 .headerLeft {
   display: flex;
   align-items: center;
