+++ src/ui/src/components/layout/Dashboard.module.css
@@ -1,7 +1,21 @@
 .dashboard {
   height: 100%;
-  overflow-y: auto;
-  padding: 28px 32px;
+  display: flex;
+  flex-direction: column;
+}
+
+.toolbar {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 10px 20px;
+  background: var(--chat-header-bg);
+  border-bottom: 1px solid var(--divider);
+  flex-shrink: 0;
+}
+
+[data-platform="mac"] .toolbar {
+  padding-left: 78px;
 }
 
 .header {
@@ -10,7 +24,6 @@
   color: var(--text-dim);
   text-transform: uppercase;
   letter-spacing: 0.08em;
-  margin-bottom: 20px;
   display: flex;
   align-items: center;
   gap: 8px;
@@ -24,7 +37,11 @@
 .grid {
   display: grid;
   grid-template-columns: repeat(auto-fill, minmax(320px, 1fr));
+  align-content: start;
   gap: 12px;
+  padding: 28px 32px;
+  overflow-y: auto;
+  flex: 1;
 }
 
 /* Workspace card */
